// src/Models/macro_counter.ts
export interface Macro_Counter{
    calories: number;
    fat: number;
    protein: number;
    carbs: number;
    noServings: number;
    saturated_fat: number;
    trans_fatty_acid: number;
    polyunsaturated_fat: number;
    monounsaturated_fat: number;
    cholesterol: number;
    sodium: number;
    dietary_fiber: number;
    sugars: number;
    vitamin_a: number;
    vitamin_c: number;
    calcium: number;
    iron: number;
    date: string;
}